+++ lib/seigen_watchdog/monitor.rb
@@ -7,12 +7,13 @@ module SeigenWatchdog
   class Monitor
     # @rbs @check_interval: Numeric?
     # @rbs @killer: Killers::Base
-    # @rbs @limiters: Hash[Symbol | String, Limiters::Base]
+    # @rbs @limiters: Hash[Symbol, Limiters::Base]
     # @rbs @logger: Logger?
     # @rbs @on_exception: Proc?
     # @rbs @before_kill: Proc?
     # @rbs @checks: Integer
     # @rbs @last_check_time: Float? | nil
+    # @rbs @time_killed: Float? | nil
     # @rbs @thread: Thread? | nil
     # @rbs @running: bool
     # @rbs @mutex: Thread::Mutex
@@ -26,7 +27,7 @@ module SeigenWatchdog
     # The killer to invoke when a limit is exceeded
     # @rbs killer: Killers::Base
     # Hash of limiters to check
-    # @rbs limiters: Hash[Symbol | String, Limiters::Base]
+    # @rbs limiters: Hash[Symbol, Limiters::Base]
     # Optional logger for debugging
     # @rbs logger: Logger?
     # Optional callback when an exception occurs
@@ -37,12 +38,13 @@ module SeigenWatchdog
     def initialize(check_interval:, killer:, limiters:, logger: nil, on_exception: nil, before_kill: nil)
       @check_interval = check_interval
       @killer = killer
-      @limiters = limiters
+      @limiters = limiters.transform_keys(&:to_sym)
       @logger = logger
       @on_exception = on_exception
       @before_kill = before_kill
       @checks = 0
       @last_check_time = nil
+      @time_killed = nil
       @thread = nil
       @running = false
       @mutex = Mutex.new
@@ -65,11 +67,14 @@ module SeigenWatchdog
       increment_checks
       log_debug("Performing check ##{@checks}")
 
-      exceeded_limiter = @limiters.values.find(&:exceeded?)
-      if exceeded_limiter
-        log_info("Limit exceeded: #{exceeded_limiter.class.name}, invoking killer")
-        run_before_kill(exceeded_limiter)
-        @killer.kill!
+      name, limiter = @limiters.find { |_k, v| v.exceeded? }
+      if limiter
+        if killed?
+          log_debug("Limit exceeded but killer already invoked #{seconds_since_killed} seconds ago")
+        else
+          log_info("Limit exceeded: #{name}, invoking killer")
+          perform_kill(name, limiter)
+        end
         true
       else
         log_debug('All limiters within bounds')
@@ -87,6 +92,21 @@ module SeigenWatchdog
       return nil if @last_check_time.nil?
 
       Process.clock_gettime(Process::CLOCK_MONOTONIC) - @last_check_time
+    end
+
+    # Returns true if the killer has been invoked
+    # @rbs return: bool
+    def killed?
+      !@time_killed.nil?
+    end
+
+    # Returns the number of seconds since the killer was invoked
+    # Seconds since killer was invoked, or nil if killer has not been invoked
+    # @rbs return: Float?
+    def seconds_since_killed
+      return nil if @time_killed.nil?
+
+      Process.clock_gettime(Process::CLOCK_MONOTONIC) - @time_killed
     end
 
     # Stops the background thread if running
@@ -114,7 +134,7 @@ module SeigenWatchdog
     # @rbs name: Symbol | String
     # @rbs return: Limiters::Base
     def limiter(name)
-      @limiters.fetch(name)
+      @limiters.fetch(name.to_sym)
     end
 
     # Returns a hash of all limiters
@@ -127,10 +147,20 @@ module SeigenWatchdog
 
     private
 
-    # @rbs exceeded_limiter: Limiters::Base
+    # @rbs name: Symbol
+    # @rbs limiter: Limiters::Base
     # @rbs return: void
-    def run_before_kill(exceeded_limiter)
-      @before_kill&.call(exceeded_limiter)
+    def perform_kill(name, limiter)
+      run_before_kill(name, limiter)
+      @killer.kill!
+      @time_killed = Process.clock_gettime(Process::CLOCK_MONOTONIC)
+    end
+
+    # @rbs name: Symbol
+    # @rbs limiter: Limiters::Base
+    # @rbs return: void
+    def run_before_kill(name, limiter)
+      @before_kill&.call(name, limiter)
     rescue StandardError => e
       handle_exception(e)
     end